fix warnings on Mac in src/views
Fix these class of warnings:
- unused functions
- unused locals
- sign mismatch
- missing function prototypes
- missing newline at end of file
- 64 to 32 bit truncation

The changes prefer to link in dead code in the debug build
with 'if (false)' than to comment it out, but trivial cases
are commented out or sometimes deleted if it appears to be
a copy/paste error.
Review URL: https://codereview.appspot.com/6300045

git-svn-id: http://skia.googlecode.com/svn/trunk@4176 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/views/mac/SkOptionsTableView.mm
@@ -50,7 +50,7 @@
 
 - (void)registerMenus:(const SkTDArray<SkOSMenu*>*)menus {
     fMenus = menus;
-    for (NSUInteger i = 0; i < fMenus->count(); ++i) {
+    for (int i = 0; i < fMenus->count(); ++i) {
         [self loadMenu:(*fMenus)[i]];
     }
 }
